move some classes to libmongocrypt
