+++ tests/MongoDB.Driver.Tests/Encryption/EncryptOptionsTests.cs
@@ -16,6 +16,7 @@
 using System;
 using FluentAssertions;
 using MongoDB.Driver.Encryption;
+using MongoDB.Libmongocrypt;
 using Xunit;
 
 namespace MongoDB.Driver.Tests.Encryption
